requirements.txt
accelerate 
albumentations
evaluate
flask
ipykernel
ipywidgets
jiwer
matplotlib
numpy
opencv-python
pandas
pillow
pynvml
seaborn
scipy
transformers==4.30.2
tqdm
ultralytics
--extra-index-url https://download.pytorch.org/whl/cu118
torch
--extra-index-url https://download.pytorch.org/whl/cu118
torchvision
--extra-index-url https://download.pytorch.org/whl/cu118
torchaudio